# specs/003-api-project-git/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ETC明細APIシステム
  description: |
    ETC明細の自動取得、データ管理、およびDtako連携機能を提供するAPIシステム

    ## 主な機能
    - **ETC明細ダウンロード**: 複数アカウントに対応した同期・非同期のETC明細データ取得
    - **ハッシュベースインポート**: 重複検出と変更追跡機能を備えたCSVインポート
    - **マッピング管理**: ETC明細とDtako行の紐付け管理
    - **データ保存・処理**: ローカルデータベースへの保存とバッチ処理

    ## 認証
    このAPIは現在認証なしで動作しますが、本番環境では適切な認証機構の実装を推奨します。

    ## エラーハンドリング
    すべてのエンドポイントで標準的なHTTPステータスコードとJSONエラーレスポンスを使用します。

  version: 0.0.16
  contact:
    name: ETC明細APIサポート
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 開発環境
  - url: https://etc-api.example.com
    description: 本番環境
paths:
...
  /api/etc/parse-csv:
    post:
      summary: CSVファイルパース
      description: アップロードされたCSVファイルをパース
      tags:
        - Parse
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: パース対象のCSVファイル
      responses:
        '200':
          description: パース成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  status:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # データ処理・保存
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: エラーメッセージ
        details:
          type: string
          description: 詳細情報
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: エラー発生時刻
  responses:
    BadRequest:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid request parameters"
...
    InternalServerError:
      description: 内部サーバーエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal server error"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API認証キー（将来実装予定）
security: []